# src/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Essa é a API de Controle de Caixa, que pode ser utilizada junto com a aplicação para controle de caixa de pequenos négocios",
        "version": "1.0.1",
        "title": "Controle de Caixa API",
        "termsOfService": "http://example.com/terms/",
        "contact": {
            "email": "carleonepsouza@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "servers": [
        {
            "url": "https://app-loja.azurewebsites.net/v1/api"
        }
    ],
    "host": "https://app-loja.azurewebsites.net/v1/api",
...
    "basePath": "/v1",
...
    "schemes": [
        "https",
        "http"
    ],
    "paths": {
...
        "/categories": {
            "post": {
                "tags": [
                    "products"
                ],
                "summary": "Register Product Category",
                "description": "This route will be responsible for register a product category",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Category"
                            },
                            "examples": {
                                "category": {
                                    "value": {
                                        "name": "Bebidas",
                                        "description": "Bebidas",
                                        "status": true
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "500": {
                        "description": "Sorry, Internal Server Error!"
                    },
                    "400": {
                        "description": " Bad Request!"
                    },
                    "401": {
                        "description": "Unauthorized!"
                    },
                    "403": {
                        "description": "Forbidden!"
                    },
                    "404": {
                        "description": "Not Found!"
                    },
                    "409": {
                        "description": "Conflit!"
                    },
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "$ref": "#/components/schemas/Category"
                                }
                            }
                        }
                    }
                }
            },
            "get": {
                "tags": [
                    "products"
                ],
                "summary": "Return a list of categories",
                "description": "This route will be responsible for return a list of categories",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "responses": {
                    "500": {
                        "description": "Sorry, Internal Server Error!"
                    },
                    "400": {
                        "description": "Bad Request!"
                    },
                    "401": {
                        "description": "Unauthorized!"
                    },
                    "403": {
                        "description": "Forbidden!"
                    },
                    "404": {
                        "description": "Not Found!"
                    },
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "$ref": "#/components/schemas/Category"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/categories/{id}": {
            "get": {
                "tags": [
                    "products"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Category ID",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    }
                ],
                "summary": "Return a specific product category",
                "description": "This route will be responsible for return a product category by ID",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "responses": {
                    "500": {
                        "description": "Sorry, Internal Server Error!"
                    },
                    "400": {
                        "description": "Bad Request!"
                    },
                    "401": {
                        "description": "Unauthorized!"
                    },
                    "403": {
                        "description": "Forbidden!"
                    },
                    "404": {
                        "description": "Not Found!"
                    },
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "$ref": "#/components/schemas/Category"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "products"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Category ID",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    }
                ],
                "summary": "Update a specific product category",
                "description": "This route will be responsible for update a product category by ID",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Category"
                            },
                            "examples": {
                                "category": {
                                    "value": {
                                        "name": "Bebidas",
                                        "description": "Bebidas",
                                        "status": true
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "500": {
                        "description": "Sorry, Internal Server Error!"
                    },
                    "400": {
                        "description": " Bad Request!"
                    },
                    "401": {
                        "description": "Unauthorized!"
                    },
                    "403": {
                        "description": "Forbidden!"
                    },
                    "404": {
                        "description": "Not Found!"
                    },
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "$ref": "#/components/schemas/Category"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "products"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Category ID",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    }
                ],
                "summary": "Delete a specific product category",
                "description": "This route will be responsible for delete a product category by ID",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "responses": {
                    "500": {
                        "description": "Sorry, Internal Server Error!"
                    },
                    "400": {
                        "description": "Bad Request!"
                    },
                    "401": {
                        "description": "Unauthorized!"
                    },
                    "403": {
                        "description": "Forbidden!"
                    },
                    "404": {
                        "description": "Not Found!"
                    },
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
        "/cashies": {
            "post": {
                "tags": [
                    "orders"
                ],
                "summary": "Register the cashie of the day",
                "description": "This route will be responsible for register a cashie of the day",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Cashie"
                            },
                            "examples": {
                                "cashie": {
                                    "value": {
                                        "user": "640a0d0a9b99632ae8b84ab0",
                                        "orders": {
                                            "type":"array",
                                            "$ref":"#/components/schemas/Order"
                                        },
                                        "valorAbertura": 50.0,
                                        "valorFechamento":50.0,
                                        "status": true,
                                        "criadoEm":{
                                            "type":"date"
                                        },
                                        "fechadoEm": {
                                            "type":"date"
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "500": {
                        "description": "Sorry, Internal Server Error!"
                    },
                    "400": {
                        "description": "Bad Request!"
                    },
                    "401": {
                        "description": "Unauthorized!"
                    },
                    "403": {
                        "description": "Forbidden!"
                    },
                    "404": {
                        "description": "Not Found!"
                    },
                    "409": {
                        "description": "Conflit!"
                    },
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "$ref": "#/components/schemas/Cashie"
                                }
                            }
                        }
                    }
                }
            },
            "get": {
                "tags": [
                    "orders"
                ],
                "summary": "Return a list of cashies",
                "description": "This route will be responsible for return a list of cashies",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "responses": {
                    "500": {
                        "description": "Sorry, Internal Server Error!"
                    },
                    "400": {
                        "description": "Bad Request!"
                    },
                    "401": {
                        "description": "Unauthorized!"
                    },
                    "403": {
                        "description": "Forbidden!"
                    },
                    "404": {
                        "description": "Not Found!"
                    },
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "$ref": "#/components/schemas/Cashie"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/cashies/{id}": {
            "get": {
                "tags": [
                    "orders"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Cashie ID",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    }
                ],
                "summary": "Return a specific cashie",
                "description": "This route will be responsible for return a cashie by ID",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "responses": {
                    "500": {
                        "description": "Sorry, Internal Server Error!"
                    },
                    "400": {
                        "description": "Bad Request!"
                    },
                    "401": {
                        "description": "Unauthorized!"
                    },
                    "403": {
                        "description": "Forbidden!"
                    },
                    "404": {
                        "description": "Not Found!"
                    },
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "$ref": "#/components/schemas/Cashie"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "orders"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Cashie ID",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    }
                ],
                "summary": "Delete a specific cashie",
                "description": "This route will be responsible for delete a cashie by ID",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "responses": {
                    "500": {
                        "description": "Sorry, Internal Server Error!"
                    },
                    "400": {
                        "description": "Bad Request!"
                    },
                    "401": {
                        "description": "Unauthorized!"
                    },
                    "403": {
                        "description": "Forbidden!"
                    },
                    "404": {
                        "description": "Not Found!"
                    },
                    "200": {
                        "description": "OK"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Product": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "category": {
                        "$ref": "#/components/schemas/Category"
                    },
                    "quantity": {
                        "type": "number"
                    },
                    "value": {
                        "type": "number"
                    },
                    "classification": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "id": {
                        "type": "string"
                    },
                    "volume": {
                        "type": "number"
                    }
                }
            },
            "Category": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "status": {
                        "type": "boolean"
                    }
                }
            },
            "User": {
                "type": "object",
                "properties": {
                    "fullName": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "phone": {
                        "type": "number"
                    },
                    "cpfCnpj": {
                        "type": "number"
                    },
                    "password": {
                        "type": "string"
                    },
                    "apiKey": {
                        "type": "string"
                    },
                    "address": {
                        "$ref": "#/components/schemas/Address"
                    },
                    "profile": {
                        "$ref": "#/components/schemas/Profile"
                    },
                    "status": {
                        "type": "boolean"
                    }
                }
            },
            "Address": {
                "type": "object",
                "properties": {
                    "street": {
                        "type": "string"
                    },
                    "neighborhood": {
                        "type": "string"
                    },
                    "zipCode": {
                        "type": "number"
                    },
                    "status": {
                        "type": "boolean"
                    }
                }
            },
            "Profile": {
                "type": "object",
                "properties": {
                    "role": {
                        "type": "string"
                    },
                    "status": {
                        "type": "boolean"
                    }
                }
            },
            "Order": {
                "type": "object",
                "properties": {
                    "nvenda": {
                        "type": "number"
                    },
                    "total": {
                        "type": "number"
                    },
                    "formaPagamnto": {
                        "type": "string"
                    },
                    "troco": {
                        "type": "number"
                    },
                    "valorPago": {
                        "type": "number"
                    },
                    "produtos": {
                        "type":"array",
                        "$ref": "#/components/schemas/Product"
                    },
                    "user": {
                        "$ref": "#/components/schemas/User"
                    },
                    "status": {
                        "type": "boolean"
                    }
                }
            },
            "Cashie": {
                "type": "object",
                "properties": {
                    "valorFechamento": {
                        "type": "number"
                    },
                    "valorAbertura": {
                        "type": "number"
                    },
                    "criadoEm": {
                        "type": "date"
                    },
                    "fechadoEm": {
                        "type": "date"
                    },
                    "orders": {
                        "type": "array",
                        "$ref": "#/components/schemas/Order"
                    },
                    "user": {
                        "$ref": "#/components/schemas/User"
                    },
                    "status": {
                        "type": "boolean"
                    }
                }
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "Bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}